- lots on byte -> int stuff related to text output - renamed the weapons ranged functions

+++ Classes/Gbl.cs
@@ -214,8 +214,8 @@ namespace Classes
         public static byte byte_1C01A;
         public static byte byte_1C01B;
         public static byte[] byte_1C8C2 = new byte[8];
-        public static byte textXCol; // byte_1C8CA
-        public static byte textYCol; // byte_1C8CB
+        public static int textXCol; // byte_1C8CA
+        public static int textYCol; // byte_1C8CB
         public static byte byte_1D1BB;
         public static byte byte_1D1C0;
         public static byte byte_1D1C4;
